feat: change location row and col types to u16
usize is a big headache to calculate the heuristc for.

consider if
g - goal
l - location
x and y correspond to rows and columns

then,
(gx - lx)       + (gy - ly)

compiles down to
(usize - usize) + (usize - usize)
which is simply out of the question

unless you convert usize to i128 so it becomes

(i128::from(usize) - i128::from(usize)) + (i128::from(usize) -
i128::from(usize))

that's way too annoying, so i'll just simply restrict the number
of rows and columns to fit into a u16.

the range of a u16 is 65535, which is big enough for any map.

rust is very, very explicit.